feat: split mfa, clean up animation
